extract processor to its own file + check if we have a template file available to rewrite it
